+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - General tab added to the remote detail page. It contains general information about the remote, such as the remote disk info, jobs, and operations.
 - Remote Clone feature added. Under the remote detail ellipsis button (Clones a remote with settings to new remote.).
 - Rclone pid watcher feature added with instant stop Rclone process functionality. Also listens for changes in the rclone process state and updates the UI accordingly. You can find it in `About RClone Manager > About Rclone`  (I see the core/pid rcd command and I want to make something for it. IDK why but I did it.)
+- Detecting the metered connection and showing a warning banner (For now, it is only show the warning banner.).
 
 ### Changed
 - Mount path selection not forced to select a path from the file browser. You can also type the path manually but it will be validated.